Separate out the entity-capturing into its own shared MonoBehaviour
