+++ Assets/Scripts/ECS/Hybrid/HybridEntity.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4043c14c8962bdd428034364d02ff179
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
